eks_automation/requirements.txt
pylint
black
pre-commit

loguru
Jinja2
requests
pygithub
boto3
